computer-systems/lectures/mongo_db/requirements.txt
colorama==0.4.6
dnspython==2.6.1
numpy==2.0.0
pandas==2.2.2
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
